LayoutTests/imported/w3c/web-platform-tests/trusted-types/block-string-assignment-to-Document-write-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.

CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'win.trustedTypes.createPolicy')

Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'win.trustedTypes.createPolicy')
